Add files via upload (#7)
* Add files via upload

I have a mutation function that should work for all three. If you want to do Evolution Strategies you just need to send in an extra Chromosome that relates to the evolution strategy matrix. For the other two, just pass in null. 
I am assuming that the adjacencyMatrix has a diagonal of only 1's, the top half of the diagonal is random weights and the bottom half is 0. If this is not the case let me know and I can adjust. I also changed the Chromosome to make adjacencyMatrix public so I could directly access the weights. If this is a problem let me know. 
Let me know if there are any issues or problems with the code or for anything you don't understand.

* Format